pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plume"
description = "A hypopycnal sediment-carrying plume entering the ocean."
authors = [
  {email = "mcflugen@gmail.com"},
  {name = "Eric Hutton"}
]
maintainers = [
  {email = "mcflugen@gmail.com"},
  {name = "Eric Hutton"}
]
keywords = [
  "bmi",
  "landlab",
  "sediment",
  "hypopycnal plume",
]
license = {file = "LICENSE.rst"}
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Cython",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: Implementation :: CPython",
  "Topic :: Scientific/Engineering :: Physics",
]
requires-python = ">=3.8"
dependencies = [
  "landlab>=2.0.0b7",
  "numpy",
  "pyyaml",
  "rich-click",
  "scipy",
]
dynamic = ["readme", "version"]

[project.urls]
homepage = "https://github.com/csdms"
documentation = "https://github.com/mcflugen/plume/blob/main/README.rst"
repository = "https://github.com/mflugen/plume"
changelog = "https://github.com/mcflugen/plume/blob/main/CHANGES.rst"

[project.optional-dependencies]
dev = [
  "black",
  "flake8",
  "flake8-bugbear",
  "isort",
  "pre-commit",
]
testing = [
  "coveralls",
  "hypothesis",
  "pytest",
  "pytest-cov",
  "pytest-datadir",
  "pytest-xdist",
  "pyyaml",
]


[project.scripts]
plume = "plume.cli:plume"


[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.dynamic]
readme = {file = ["README.rst", "AUTHORS.rst", "CHANGES.rst"]}
version = {attr = "plume._version.__version__"}

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["plume", "tests"]
norecursedirs = [".*", "*.egg*", "build", "dist", "examples"]
addopts = """
  --ignore setup.py
  --tb native
  --strict
  --durations 16
  --doctest-modules
  -vvv
"""
doctest_optionflags = [
  "NORMALIZE_WHITESPACE",
  "IGNORE_EXCEPTION_DETAIL",
  "ALLOW_UNICODE"
]

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true
line_length = 88

[tool.towncrier]
directory = "news"
package = "plume"
filename = "CHANGES.rst"
single_file = true
underlines = "-`^"
issue_format = "`#{issue} <https://github.com/mcflugen/plume/issues/{issue}>`_"
title_format = "{version} ({project_date})"

[[tool.towncrier.type]]
directory = "feature"
name = "New Features"
showcontent = true

[[tool.towncrier.type]]
directory = "bugfix"
name = "Bug Fixes"
showcontent = true

[[tool.towncrier.type]]
directory = "docs"
name = "Documentation Enhancements"
showcontent = true

[[tool.towncrier.type]]
directory = "misc"
name = "Other Changes and Additions"
showcontent = true
